backend/requirements.txt
fastapi
uvicorn
python-dotenv
groq
langgraph
langchain
langchain-core
langchain-groq
opencv-python
ultralytics
websockets
numpy